src/ace_time/testing: Cast (int32_t) to (long) to match '%ld' printf() format

+++ src/ace_time/testing/BasicTransitionTest.h
@@ -109,7 +109,7 @@ class BasicTransitionTest: public aunit::TestOnce {
       logging::printf(
           "* failed %s: index=%d eps=%ld "
           "%04d-%02d-%02dT%02d:%02d:%02d: ",
-          tag, i, item.epochSeconds,
+          tag, i, (long) item.epochSeconds,
           item.year, item.month, item.day,
           item.hour, item.minute, item.second);
     }
